Read Me: i must say refreshing!
It is nice to find someone that can be honest and not out for theirselfs.  Like most of these people who are out for your money and you tell them no and they still bother you for four to five more days. I've gotten to the point if I don't have a specific  ring  tone for you I won't  answer  my phone nor do I pay attention to texts that much. Honesty we need more of that.